requirements.txt
boto3==1.35.54
jinja2==3.1.4
pillow==11.0.0
pyqt6==6.7.1
pytest==8.3.4
